Refactoring: Extract AssignmentFinder
